+++ lesson9/src/main/kotlin/demo/application/controller/CreditApplication.kt
@@ -1,0 +1,10 @@
+package demo.application.controller
+
+import java.time.LocalDateTime
+
+data class CreditApplication(
+    val createdAt: LocalDateTime,
+    val totalAmount: Long,
+    val monthlyPayment: Long,
+    val user: User,
+)
