+++ lesson9/src/main/kotlin/demo/application/controller/Loan.kt
@@ -1,0 +1,10 @@
+package demo.application.controller
+
+import java.time.LocalDateTime
+
+data class Loan(
+    val createAt: LocalDateTime,
+    val loanTerm: LocalDateTime,
+    val isClose: Boolean,
+    val monthlyPayment: Long,
+)
