+++ lesson9/src/main/kotlin/demo/application/controller/ScoreController.kt
@@ -1,40 +1,13 @@
 package demo.application.controller
-
-import org.springframework.web.bind.annotation.PostMapping
-import org.springframework.web.bind.annotation.RequestBody
+import org.springframework.web.bind.annotation.GetMapping
+import org.springframework.web.bind.annotation.RequestParam
 import org.springframework.web.bind.annotation.RestController
 import java.time.LocalDateTime
-
 @RestController
 class ScoreController(
     val logicService: LogicService,
 ) {
-
-
-    @PostMapping("/score")
-    fun simpleScore(@RequestBody creditApplication: CreditApplication) = logicService.simpleScore(creditApplication)
-
+    @GetMapping("score")
+    fun simpleScore(@RequestParam creditApplication: CreditApplication) =
+        logicService.simpleScore(creditApplication)
 }
-
-data class User(
-    val age: Int,
-    val name: String,
-    val sex: Int, // ????
-    val income: Long,
-    val loans: List<Loan>,
-) {
-
-    data class Loan(
-        val createdAt: LocalDateTime,
-        val isClosed: Boolean,
-        val monthlyPayment: Long,
-    )
-
-}
-
-data class CreditApplication(
-    val createdAt: LocalDateTime,
-    val totalAmount: Long,
-    val monthlyPayment: Long,
-    val user: User,
-)
