+++ lesson9/src/main/kotlin/demo/application/controller/User.kt
@@ -1,0 +1,10 @@
+package demo.application.controller
+
+data class User(
+    val age: Int,
+    val name: String,
+    val sex: Int,
+    val creditScore: Int,
+    val income: Long,
+    val loans: List<demo.application.controller.Loan>,
+)
